image-builder/scripts/requirements.txt
requests==2.28.1
python-dotenv==0.20.0
PyGithub==1.55